Refactor elasticsearch-service to the new project structure and manual test ess and ae with mds new changes
