+++ elasticsearch-service/src/elasticsearchModels/dashboardSeedModel.ts
@@ -1,5 +1,5 @@
-import { IAggregation } from "../../../common-service/src/models/aggregationModel";
-import { IJob } from "../../../common-service/src/models/jobModel";
+import { IAggregation } from "../../../mongodb-service/src/models/aggregationModel";
+import { IJob } from "../../../mongodb-service/src/models/jobModel";
 
 export interface IDashboardSeed {
     job: IJob;
