+++ mongodb-service/.env
@@ -1,4 +1,4 @@
-MONGODB_URI=mongodb://root:passw0rd@localhost:27017/metadata?authSource=admin
+MONGODB_URI=mongodb://root:passw0rd@localhost:27017/gdatool?authSource=admin
 ENABLE_CORS=false
 ENABLE_SWAGGER=true
 JWT_SECRET=gdatool-secret
